+++ package.xml
@@ -8,7 +8,6 @@
   <license>Apache 2.0</license>
   <author email="ilario.azzollini92@gmail.com">Ilario Azzollini</author>
 
-  
   <depend>geometry_msgs</depend>
   <depend>irobot_create_msgs</depend>
   <depend>nav_msgs</depend>
